+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [2.6.0]
+
+### Changed
+
+- Refactored `\FQL\Conditions\BaseConditionGroup` and make it abstract and implements concrete condition
+groups `WhereConditionGroup` and `HavingConditionGroup` and `IfStatementConditionGroup` and `CaseStatementConditionGroup`
+
+### Added
+- `IF` and `CASE-WHEN` supporting multiple conditions as `WHERE` or `HAVING` do.
+- New operator `REGEXP` and `NOT REGEXP`
+
 ## [2.5.3]
 
 ### Changed